first networking test

+++ server/game/input.go
@@ -1,6 +1,6 @@
 package game
 
 type Input struct {
-	Ax, Ay float32
+	Ax, Ay float64
 	Boost  bool
 }
