+++ server/protocol/server.go
@@ -6,15 +6,15 @@ type Welcome struct {
 }
 
 type State struct {
-	Tick    uint64           `json:"tick"`
+	Tick    int              `json:"tick"`
 	Players []PlayerSnapshot `json:"players"`
 }
 
 type PlayerSnapshot struct {
 	ID string  `json:"id"`
-	X  float32 `json:"x"`
-	Y  float32 `json:"y"`
-	A  float32 `json:"a,omitempty"` // optional angle
+	X  float64 `json:"x"`
+	Y  float64 `json:"y"`
+	A  float64 `json:"a,omitempty"` // optional angle
 }
 
 type Error struct {
